requirements.txt
# requirements.txt
fastapi>=0.68.0
uvicorn>=0.15.0
python-dotenv>=1.0.0
cryptography>=41.0.7
pybit>=2.6.0
numpy>=1.21.0
TA-Lib==0.4.24